move parMapN to its own file. also clean up some suspended lambas in tests
